+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 New Features
 ^^^^^^^^^^^^
++ Creation of new data classes for handling aperture, photometry, and catalog data in a more consistent way. [#125]
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -11,6 +12,7 @@ Other Changes and Additions
 
 Bug Fixes
 ^^^^^^^^^
++ Fixed dependence on non-release version of astrowidgets for overwrite capability on output images. [#108]
 
 1.3.9 (2023-06-16)
 ------------------